requirements.txt
fastapi
langchain
langchain-google-genai
redis
pypdf2
python-multipart
uvicorn
faker
python-dotenv
requests
